+++ JournalMonitor/JournalMonitor.cs
@@ -2077,7 +2077,7 @@ namespace EddiJournalMonitor
                                 data.TryGetValue("Cost", out val);
                                 long price = (long)val;
 
-                                events.Add(new ShipRefuelledEvent(timestamp, "Market", price, amount, null) { raw = line });
+                                events.Add(new ShipRefuelledEvent(timestamp, "Market", price, amount, null, false) { raw = line });
                                 handled = true;
                                 break;
                             }
@@ -2088,7 +2088,7 @@ namespace EddiJournalMonitor
                                 data.TryGetValue("Cost", out val);
                                 long price = (long)val;
 
-                                events.Add(new ShipRefuelledEvent(timestamp, "Market", price, amount, null) { raw = line });
+                                events.Add(new ShipRefuelledEvent(timestamp, "Market", price, amount, null, true) { raw = line });
                                 handled = true;
                                 break;
                             }
@@ -2097,7 +2097,9 @@ namespace EddiJournalMonitor
                                 decimal amount = JsonParsing.getDecimal(data, "Scooped");
                                 decimal total = JsonParsing.getDecimal(data, "Total");
 
-                                events.Add(new ShipRefuelledEvent(timestamp, "Scoop", null, amount, total) { raw = line });
+                                events.Add(new ShipRefuelledEvent(timestamp, "Scoop", null, amount, total, 
+                                    EDDI.Instance.CurrentShip == null ? false : EDDI.Instance.CurrentShip.fueltanktotalcapacity == null ? false : 
+                                    Math.Round(total) == EDDI.Instance.CurrentShip.fueltanktotalcapacity) { raw = line });
                                 handled = true;
                                 break;
                             }