+++ Events/SearchAndRescueEvent.cs
@@ -1,4 +1,5 @@
 using EddiDataDefinitions;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 
@@ -13,24 +14,26 @@ namespace EddiEvents
 
         static SearchAndRescueEvent()
         {
-    	    VARIABLES.Add("commodity", "The name of the item recovered");
+    	    VARIABLES.Add("commodity", "The commodity (object) recovered");
+            VARIABLES.Add("localizedcommodityname", "The localized name of the commodity recovered");
             VARIABLES.Add("amount", "The amount of the item recovered");
             VARIABLES.Add("reward", "The monetary reward for completing the search and rescue");
         }
-
-        public string commodity => commodityDefinition?.localizedName ?? "unknown commodity";
 
         public int? amount { get; }
 
         public long reward { get; }
 
-        public CommodityDefinition commodityDefinition { get; }
+        public string localizedcommodityname => commodity.localizedName;
+
+        [JsonProperty("commodity")]
+        public CommodityDefinition commodity { get; }
 
         public SearchAndRescueEvent(DateTime timestamp, CommodityDefinition commodity, int? amount, long reward) : base(timestamp, NAME)
         {
             this.amount = amount;
             this.reward = reward;
-            this.commodityDefinition = commodity;
+            this.commodity = commodity;
         }
     }
 }