+++ EDDI/ChangeLog.md
@@ -5,12 +5,15 @@ Full details of the variables available for each noted event, and VoiceAttack in
 ### 3.0.1-b5
   * Core
     * Updated system definition to include new variable `lastVisitSeconds`. 
+    * The Search and rescue event was having its `commodity` property set to just the commodity name, rather than the commodity definition object that scripts expect. Fixed.
   * Speech responder
     * Updated `Commodity sales check` script to make use of `CommodityMarketDetails()` function.
     * Fixed a bug that had made `{ship.role}` inaccessible via scripts.
     * Updated `FSD engaged` script to correct a bug that was preventing sub-function `System report` from ever running.
   * Added the following Cottle function, documented in [the SpeechResponder documentation](https://github.com/EDCD/EDDI/blob/master/SpeechResponder/Help.md):
      * `CommodityMarketDetails()` for retrieving market information about commodities.
+  * VoiceAttack
+    * The Search and rescue event has a new variable `{TXT:localizedcommodityname}` as the `commodity` varaible is now an object.
 
 ### 3.0.1-b4
   * Core